extract card storage
