+++ infrastructure/storage/card-codec.ts
@@ -1,0 +1,37 @@
+import type { Card as FsrsCard } from 'ts-fsrs';
+import type { Card } from '@/domain/cards';
+
+export interface StoredCard extends Omit<Card, 'createdAt' | 'fsrs' | 'domain'> {
+  domain?: Card['domain'];
+  createdAt: number;
+  fsrs: Omit<FsrsCard, 'due' | 'last_review'> & {
+    due: number;
+    last_review?: number;
+  };
+}
+
+export function serializeCard(card: Card): StoredCard {
+  return {
+    ...card,
+    createdAt: card.createdAt.getTime(),
+    fsrs: {
+      ...card.fsrs,
+      due: card.fsrs.due.getTime(),
+      last_review: card.fsrs.last_review?.getTime(),
+    },
+  };
+}
+
+export function deserializeCard(stored: StoredCard): Card {
+  const { due, last_review, ...rest } = stored.fsrs;
+  return {
+    ...stored,
+    domain: stored.domain ?? 'leetcode.com',
+    createdAt: new Date(stored.createdAt),
+    fsrs: {
+      ...rest,
+      due: new Date(due),
+      last_review: last_review ? new Date(last_review) : undefined,
+    },
+  };
+}
