+++ services/__tests__/cards.test.ts
@@ -1,24 +1,13 @@
-import { createEmptyCard, State as FsrsState, Rating } from 'ts-fsrs';
+import { State as FsrsState, Rating } from 'ts-fsrs';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 import { fakeBrowser } from 'wxt/testing/fake-browser';
 import { storage } from 'wxt/utils/storage';
-import type { Card } from '@/domain/cards';
 import type { DailyStats } from '@/domain/stats';
+import type { StoredCard } from '@/infrastructure/storage/card-codec';
 import { STORAGE_KEYS } from '@/infrastructure/storage/storage-keys';
 import { requireDefined } from '@/test/utils/assertions';
 import { buildSettings } from '@/test/utils/settings-mocks';
-import {
-  addCard,
-  delayCard,
-  deserializeCard,
-  getAllCards,
-  getReviewQueue,
-  rateCard,
-  removeCard,
-  type StoredCard,
-  serializeCard,
-  setPauseStatus,
-} from '../cards';
+import { addCard, delayCard, getAllCards, getReviewQueue, rateCard, removeCard, setPauseStatus } from '../cards';
 import * as notesModule from '../notes';
 
 const { mockGetSettings } = vi.hoisted(() => ({ mockGetSettings: vi.fn() }));
@@ -36,135 +25,6 @@ vi.mock('../settings', () => ({
 
 beforeEach(() => {
   mockGetSettings.mockResolvedValue(buildSettings());
-});
-
-describe('Card serialization', () => {
-  describe('serializeCard', () => {
-    it('should convert Date to timestamp', () => {
-      const testDate = new Date('2024-01-15T10:30:00Z');
-      const card: Card = {
-        id: 'test-id-1',
-        slug: 'two-sum',
-        name: 'Two Sum',
-        leetcodeId: '1',
-        difficulty: 'Easy',
-        createdAt: testDate,
-        fsrs: createEmptyCard(),
-        paused: false,
-        domain: 'leetcode.com',
-      };
-
-      const serialized = serializeCard(card);
-
-      expect(serialized.slug).toBe('two-sum');
-      expect(serialized.name).toBe('Two Sum');
-      expect(serialized.createdAt).toBe(testDate.getTime());
-      expect(typeof serialized.createdAt).toBe('number');
-    });
-
-    it('should serialize FSRS card dates', () => {
-      const testDate = new Date('2024-01-15T10:30:00Z');
-      const fsrsCard = createEmptyCard();
-      fsrsCard.last_review = new Date('2024-01-14T09:00:00Z');
-
-      const card: Card = {
-        id: 'test-id-2',
-        slug: 'two-sum',
-        name: 'Two Sum',
-        leetcodeId: '1',
-        difficulty: 'Medium',
-        createdAt: testDate,
-        fsrs: fsrsCard,
-        paused: false,
-        domain: 'leetcode.com',
-      };
-
-      const serialized = serializeCard(card);
-
-      expect(typeof serialized.fsrs.due).toBe('number');
-      expect(serialized.fsrs.due).toBe(fsrsCard.due.getTime());
-      expect(typeof serialized.fsrs.last_review).toBe('number');
-      expect(serialized.fsrs.last_review).toBe(fsrsCard.last_review.getTime());
-      expect(serialized.fsrs.stability).toBe(fsrsCard.stability);
-      expect(serialized.fsrs.difficulty).toBe(fsrsCard.difficulty);
-    });
-  });
-
-  describe('deserializeCard', () => {
-    it('should convert timestamp back to Date object', () => {
-      const timestamp = new Date('2024-01-15T10:30:00Z').getTime();
-      const emptyFsrs = createEmptyCard();
-      const storedCard: StoredCard = {
-        id: 'test-id-3',
-        slug: 'merge-intervals',
-        name: 'Merge Intervals',
-        leetcodeId: '56',
-        difficulty: 'Hard',
-        createdAt: timestamp,
-        fsrs: {
-          ...emptyFsrs,
-          due: emptyFsrs.due.getTime(),
-          last_review: emptyFsrs.last_review?.getTime(),
-        },
-        paused: false,
-        domain: 'leetcode.com',
-      };
-
-      const deserialized = deserializeCard(storedCard);
-
-      expect(deserialized.slug).toBe('merge-intervals');
-      expect(deserialized.name).toBe('Merge Intervals');
-      expect(deserialized.createdAt).toBeInstanceOf(Date);
-      expect(deserialized.createdAt.getTime()).toBe(timestamp);
-    });
-
-    it('should default domain to leetcode.com when missing', () => {
-      const timestamp = new Date('2024-01-15T10:30:00Z').getTime();
-      const emptyFsrs = createEmptyCard();
-      // Simulate a pre-migration stored card that lacks the domain field
-      const storedCard = {
-        id: 'test-id-old',
-        slug: 'old-problem',
-        name: 'Old Problem',
-        leetcodeId: '100',
-        difficulty: 'Easy',
-        createdAt: timestamp,
-        fsrs: {
-          ...emptyFsrs,
-          due: emptyFsrs.due.getTime(),
-          last_review: emptyFsrs.last_review?.getTime(),
-        },
-        paused: false,
-      } as unknown as StoredCard;
-
-      const deserialized = deserializeCard(storedCard);
-
-      expect(deserialized.domain).toBe('leetcode.com');
-    });
-  });
-
-  describe('serializeCard and deserializeCard roundtrip', () => {
-    it('should maintain data integrity through serialization and deserialization', () => {
-      const originalCard: Card = {
-        id: 'test-id-4',
-        slug: 'two-pointers',
-        name: 'Two Pointers',
-        leetcodeId: '999',
-        difficulty: 'Medium',
-        createdAt: new Date(),
-        fsrs: createEmptyCard(),
-        paused: false,
-        domain: 'leetcode.com',
-      };
-
-      const serialized = serializeCard(originalCard);
-      const deserialized = deserializeCard(serialized);
-
-      expect(deserialized.slug).toBe(originalCard.slug);
-      expect(deserialized.name).toBe(originalCard.name);
-      expect(deserialized.createdAt.getTime()).toBe(originalCard.createdAt.getTime());
-    });
-  });
 });
 
 describe('addCard', () => {
@@ -306,91 +166,6 @@ describe('addCard', () => {
     expect(typeof storedCard.createdAt).toBe('number');
     expect(storedCard.slug).toBe(card.slug);
     expect(storedCard.name).toBe(card.name);
-  });
-});
-
-describe('getAllCards', () => {
-  beforeEach(() => {
-    // Reset the fake browser state before each test
-    fakeBrowser.reset();
-  });
-
-  it('should return empty array when no cards exist', async () => {
-    const cards = await getAllCards();
-    expect(cards).toEqual([]);
-  });
-
-  it('should return all cards from storage', async () => {
-    // Add multiple cards
-    await addCard({ slug: 'two-sum', name: 'Two Sum', leetcodeId: '1', difficulty: 'Easy', domain: 'leetcode.com' });
-    await addCard({
-      slug: 'valid-parentheses',
-      name: 'Valid Parentheses',
-      leetcodeId: '20',
-      difficulty: 'Medium',
-      domain: 'leetcode.com',
-    });
-    await addCard({
-      slug: 'merge-intervals',
-      name: 'Merge Intervals',
-      leetcodeId: '56',
-      difficulty: 'Hard',
-      domain: 'leetcode.com',
-    });
-
-    // Get all cards
-    const allCards = await getAllCards();
-
-    expect(allCards).toHaveLength(3);
-
-    // Check that all cards are present
-    const cardSlugs = allCards.map((c) => c.slug);
-    expect(cardSlugs).toContain('two-sum');
-    expect(cardSlugs).toContain('valid-parentheses');
-    expect(cardSlugs).toContain('merge-intervals');
-
-    // Check that cards have correct data
-    const foundCard1 = allCards.find((c) => c.slug === 'two-sum');
-    expect(foundCard1?.name).toBe('Two Sum');
-
-    const foundCard2 = allCards.find((c) => c.slug === 'valid-parentheses');
-    expect(foundCard2?.name).toBe('Valid Parentheses');
-
-    const foundCard3 = allCards.find((c) => c.slug === 'merge-intervals');
-    expect(foundCard3?.name).toBe('Merge Intervals');
-  });
-
-  it('should properly deserialize stored cards', async () => {
-    const testDate = new Date('2024-01-15T10:30:00Z');
-
-    // Manually add a serialized card to storage
-    const emptyFsrs = createEmptyCard();
-    const storedCard: StoredCard = {
-      id: 'test-id-deserialize',
-      slug: 'test-problem',
-      name: 'Test Problem',
-      leetcodeId: '999',
-      difficulty: 'Medium',
-      createdAt: testDate.getTime(),
-      fsrs: {
-        ...emptyFsrs,
-        due: emptyFsrs.due.getTime(),
-        last_review: emptyFsrs.last_review?.getTime(),
-      },
-      paused: false,
-      domain: 'leetcode.com',
-    };
-
-    await storage.setItem(STORAGE_KEYS.cards, { 'test-problem': storedCard });
-
-    // Get all cards
-    const allCards = await getAllCards();
-
-    expect(allCards).toHaveLength(1);
-    expect(allCards[0].slug).toBe('test-problem');
-    expect(allCards[0].name).toBe('Test Problem');
-    expect(allCards[0].createdAt).toBeInstanceOf(Date);
-    expect(allCards[0].createdAt.getTime()).toBe(testDate.getTime());
   });
 });
